% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{convert_genomic_to_continuous_axis}
\alias{convert_genomic_to_continuous_axis}
\title{Convert genomic coordinates to a single continuous x axis for plotting}
\usage{
convert_genomic_to_continuous_axis(dt_bins)
}
\arguments{
\item{dt_bins}{data.table holding genomic bins with columns
\itemize{
\item Chrom
\item start
\item end
}}
}
\description{
Convert genomic coordinates to a single continuous x axis for plotting
}
