% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{define_chromosome_sizes}
\alias{define_chromosome_sizes}
\title{Define chromosome sizes based on species.}
\usage{
define_chromosome_sizes(species = "Human", only_autosomes = TRUE)
}
\arguments{
\item{species}{Character name of spieces one of "Human" or "Mouse"}

\item{only_autosomes}{Logical if to include only autosomes. Default: TRUE}
}
\description{
Define chromosome sizes based on species.
}
