% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{detect_sample_mode}
\alias{detect_sample_mode}
\title{Determine if sample is matched or not}
\usage{
detect_sample_mode(results_path)
}
\arguments{
\item{results_path}{Character, path to results for a certain sample.}
}
\value{
Chacacter indicating if results are from "matched" healthy-tumor
sample pair or "single" only tumor sample. See \code{SAMPLE_MODES}
}
\description{
Determine if sample is matched or not
}
