% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_mocaseq_cna_callers}
\alias{get_mocaseq_cna_callers}
\title{Parse available CNV caller results form MoCaSeq folder}
\usage{
get_mocaseq_cna_callers(sample_id, result_dir_path = ".")
}
\arguments{
\item{sample_id}{Character sample id.}

\item{result_dir_path}{Character path to result directory. Default: "."}
}
\description{
Infer method used for copy number variation calling from MoCaSeq restults
folder.
}
