% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_mocaseq_loh_file}
\alias{get_mocaseq_loh_file}
\title{Build path to LOH variants from MoCaSeq}
\usage{
get_mocaseq_loh_file(
  sample_name,
  result_type = "variants_for_LOH",
  variant_type = "somatic"
)
}
\arguments{
\item{sample_name}{Character sample name from MoCaSeq run}

\item{variant_type}{Character variant type one of "germline" or "somatic"}
}
\description{
MoCaSeq produces loss of heterozygosity data. This function returns the name
of the file containing these variants, given a sample name and which type of
variant desired (i.e. germline or somatic).
}
