% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_mocaseq_snv_file}
\alias{get_mocaseq_snv_file}
\title{IO code for MoCaSeq}
\usage{
get_mocaseq_snv_file(
  sample_name,
  sample_type,
  mutation_caller = "Mutect2",
  postprocessing = "NoCommonSNPs",
  pipeline_version = "bash"
)
}
\arguments{
\item{sample_name}{Character, name of sample}

\item{sample_type}{Character, one of "Tumor", "Normal", "matched" indicating
type of sample}

\item{postprocessing}{Character, describing postprocessing applied to
mutations E.g. "Positions", "NoCommonSNPs"}
}
\description{
Functions for building paths, and reading or writing data related to the
MoCaSeq pipeline
Build path to SNP file in MoCaSeq results
}
\details{
MoCaSeq uses MuTect2 for mutation calling, this function creates the standard
path to the (postprocessed) SNP file given the sample name and sample type.
Please note, cancer genomics requires discerning germline and somatic
nucleotide variants. We refer to germline variants as single nucleotide
polymorphisms (SNPs) and somatic variants as single nucleotide variants
(SNVs).
}
