% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.p12.centromeres}
\alias{GRCh38.p12.centromeres}
\title{Centromeres of human genome v38p12}
\format{
\subsection{\code{GRCh38.p12.centromeres}}{

A data frame with 24 rows and 4 columns:
\describe{
\item{chr}{Chromosome name}
\item{start, end}{Chromosomal coordinates of start and end position}
\item{class}{Class of genomics region. Here: centromere}
...
}
}
}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables}
}
\usage{
GRCh38.p12.centromeres
}
\description{
A table of chromosomal coordinates of the centromeres in the human reference
genome (GRCh38.p12).
}
\keyword{datasets}
