% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GRCh38.p12.mappability}
\alias{GRCh38.p12.mappability}
\title{Regions of poor mappavility of human genome v38p12}
\format{
\subsection{\code{GRCh38.p12.mappability}}{

A data frame with 24 rows and 4 columns:
\describe{
\item{chr}{Chromosome name}
\item{start, end}{Chromosomal coordinates of start and end position}
\item{class}{Class of genomics region. Here: UmapS36}
...
}
}
}
\source{
\url{http://genome.ucsc.edu/cgi-bin/hgTables}
}
\usage{
GRCh38.p12.mappability
}
\description{
A table of chromosomal coordinates for genomic regions of poor mappability
in the human reference genome (GRCh38.p12). These regions are poorly aligned
to the reference genome and will be prone to observe artifacts in sequencing
data.
}
\keyword{datasets}
