% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_genomic_regions}
\alias{load_genomic_regions}
\title{Loads genomic ancillary regions}
\usage{
load_genomic_regions(
  species,
  region_type,
  ref_genome = NULL,
  return_granges = FALSE
)
}
\arguments{
\item{species}{Character name of species, i.e. "human" or "mouse".}

\item{region_type}{Character specifying the type of regions to return. I.e.
one of "chromosomes", "centromeres", "alt.haplotypes", or "mappability".}

\item{ref_genome}{Character name of reference genome to use for species. Per
default (NULL) GRCh38.p12 and GRCm38.p6 will be used.}

\item{return_granges}{Logical if to return results as GenomicRanges, returns
a data.table by default (FALSE).}
}
\value{
\code{data.table::data.table} or \code{GenomicRanges::GRanges}
listing genomic regions.
}
\description{
This function loads genomic regions depending on a species (and reference
genome version). Available options are chromosomes, centromeres, alternative
haplotypes, and regions of poor mappability.
}
