% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_segments}
\alias{load_segments}
\title{Load and standardize copy number segment data}
\usage{
load_segments(
  cns_file,
  cnv_method = NULL,
  species = NULL,
  sample_id = NULL,
  column_name_map = NULL,
  column_names = c("chromosome", "start", "end", "mean_log2"),
  ...
)
}
\arguments{
\item{cns_file}{Character path to copy number segment (.cns) file. This is a
tab separated file with chromosome, start, end, and mean (log2 CN) columns.}

\item{cnv_method}{Character, name of copy number calling method. This
software was developed for \code{"Copywriter"}, \code{"HMMCopy"}, and
\code{"CNVKit"}. To use data from other CNV callers, use any other value here
together with the \code{column_name_map} parameter.}

\item{species}{Character, \code{"Human"} or \code{"Mouse"}.}

\item{sample_id}{Character sample identifier.}

\item{column_name_map}{Named character vector, specifying column renaming
from input file. I.e. \code{c(foo="bar")} will rename column \code{bar} to
\code{foo}.}

\item{column_names}{Character vector, specifying column names of returned
segment data.table. Needs to match \code{names(column_name_map)}.
Assumed 1 to 4 to be chromosome, start, end and mean CN value names.
Suggested standard: \code{c("chromosome", "start", "end", "mean_log2")}.}

\item{...}{parameters passed down to data.table::fread()}
}
\value{
A \code{data.table} with columns:
\code{chr, start, end, mean_log2} (and \code{gene} for CNVKit).
}
\description{
Reads the segment file, renames columns to a (parametrized) common schema,
and standardizes chromosome names to characters of numbers (autosomes) or
letters (sex chromosomes).
This function wraps data.table::fread for reading the segment file.
}
