% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_snv_file}
\alias{load_snv_file}
\title{Lead SNV data from file using data.table}
\usage{
load_snv_file(snv_file, pipeline_version = NULL, keep_columns = NULL, ...)
}
\arguments{
\item{snv_file}{Character path to snv file.}

\item{pipeline_version}{Character indicating MoCaSeq version. One of "bash"
or "nextflow". Default: NULL (i.e. auto detect)}

\item{keep_columns}{Character vector of column names to keep.}

\item{...}{Parameters used for data.table::fread() downstream.}
}
\description{
Wrapper for loading snv data from MoCaSeq results. Standardizes column names.
}
