% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoCaHelp-package.R
\docType{package}
\name{MoCaHelp-package}
\alias{MoCaHelp}
\alias{MoCaHelp-package}
\title{MoCaHelp: Helper Code for MoCaSeq pipeline}
\description{
This package provides standard functions for downstram analyses of MoCaSeq. It includes loading and wrangling data from MoCaSeq output, getting some metadata, and loading smaller reference genome data (e.g. centromere regions).
}
\author{
\strong{Maintainer}: Marcus Wagner \email{marcus.wagner@tum.de} (\href{https://orcid.org/0009-0000-9421-6766}{ORCID})

Authors:
\itemize{
  \item Marcus Wagner \email{marcus.wagner@tum.de} (\href{https://orcid.org/0009-0000-9421-6766}{ORCID})
}

}
\keyword{internal}
