% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{MUTECT2_COLUMN_NAMES_BASH}
\alias{MUTECT2_COLUMN_NAMES_BASH}
\title{IO code for MoCaSeq}
\usage{
MUTECT2_COLUMN_NAMES_BASH
}
\description{
Functions for building paths, and reading or writing data related to the
MoCaSeq pipeline
}
