% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome.R
\name{process_count_data}
\alias{process_count_data}
\title{Process count data from LOH}
\usage{
process_count_data(chromosome_sizes, count_data = "", method = "")
}
\arguments{
\item{count_data}{Character path to count data}

\item{method}{Character}
}
\description{
Process count data from LOH
}
\note{
version 2 uses data.table::fread instead of read.table
}
