% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{SAMPLE_TYPES}
\alias{SAMPLE_TYPES}
\title{MoCaSeq pipeline metadata}
\usage{
SAMPLE_TYPES
}
\description{
Metadata on MoCaSeq pipeline, including wich samples types can be processed,
which tools are available, etc.
}
