% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{detect_mocaseq_cnv_caller}
\alias{detect_mocaseq_cnv_caller}
\title{Detect CNV caller in MoCaSeq results}
\usage{
detect_mocaseq_cnv_caller(results_path)
}
\arguments{
\item{results_path}{Character, path to results for a certain sample.}
}
\value{
Character name of CNV caller one of "CNVKit", "HMMCopy" or
"CopyWriter".
}
\description{
MoCaSeq employs CNVKit, HMMCopy and Copywriter for calling copy number
variations (CNVs). It stores any results under \code{sample_name/results/} in
a subfolder with the respective \code{tool_name}. This function exploits this
structure by searching all available tools for a CNV caller.
}
