% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{detect_mocaseq_version}
\alias{detect_mocaseq_version}
\title{Detect MoCaSeq pipeline version}
\usage{
detect_mocaseq_version(results_path)
}
\arguments{
\item{results_path}{Character path to results for a given sample. I.e. this
path should end with sample_id/results}
}
\value{
Character 'bash', or 'nextflow' indicating MoCaSeq pipeline version
}
\description{
Detect MoCaSeq pipeline version
}
