% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_mocaseq_cnv_file}
\alias{get_mocaseq_cnv_file}
\title{Build path to copy number segments file.}
\usage{
get_mocaseq_cnv_file(
  sample_name,
  sample_type,
  cnv_caller,
  result_type = "segments",
  seg_size = 20000
)
}
\arguments{
\item{sample_name}{Character, name of sample}

\item{sample_type}{Character, type of sample ("Tumor", "Normal", or
"matched")}

\item{cnv_caller}{Character, name of CNV caller ("CNVKit", "HMMCopy" or
"Copywriter")}

\item{result_type}{Character, type of CNV result ("segments", "log2RR")}

\item{seg_size}{Integer, segment size of HMMCopy output, only used in case
of \code{cnv_caller = "HMMCopy"}.}
}
\description{
Copy number segment (cns) files produced by MoCaSeq can be either from
CNVKit, HMMcopy, or CopyWriter. This function resolves possible cns files.
}
