% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_mocaseq_path}
\alias{get_mocaseq_path}
\title{Construct path for MoCaSeq result file}
\usage{
get_mocaseq_path(
  sample_name,
  sample_type,
  tool_name,
  base_path = ".",
  variant_type = "mixed",
  ignore_not_existing = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sample_name}{Character name of sample from MoCaSeq run}

\item{sample_type}{Character sample type, one of Tumor, matched or Normal}

\item{tool_name}{Character name of tool from MoCaSeq pipeline}

\item{base_path}{Character of path to directory with MoCaSeq output

TODO: expand/recycle parameters other than sample_name if multiple sample
names are provided}
}
\value{
Character path to described file, or NULL
}
\description{
Construct path for MoCaSeq result file
}
