% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_mocaseq_snv_file}
\alias{get_mocaseq_snv_file}
\title{Build path to SNP file in MoCaSeq results}
\usage{
get_mocaseq_snv_file(
  sample_name,
  sample_type,
  mutation_caller = "Mutect2",
  postprocessing = "NoCommonSNPs",
  pipeline_version = "bash"
)
}
\arguments{
\item{sample_name}{Character, name of sample}

\item{sample_type}{Character, one of "Tumor", "Normal", "matched" indicating
type of sample}

\item{mutation_caller}{Character name of method use to infer SNVs}

\item{postprocessing}{Character, describing postprocessing applied to
mutations E.g. "Positions", "NoCommonSNPs"}

\item{pipeline_version}{Character indicating MoCaSeq pipeline version used to
obtain the results found. One of "bash", or "nextflow". Default: "bash"}
}
\description{
MoCaSeq uses MuTect2 for mutation calling, this function creates the standard
path to the (postprocessed) SNP file given the sample name and sample type.
Please note, cancer genomics requires discerning germline and somatic
nucleotide variants. We refer to germline variants as single nucleotide
polymorphisms (SNPs) and somatic variants as single nucleotide variants
(SNVs).
}
