% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoCaHelp-package.R
\docType{package}
\name{MoCaHelp-package}
\alias{MoCaHelp}
\alias{MoCaHelp-package}
\title{MoCaHelp: Helper code for Mouse Cancer Sequencing (MoCaSeq) pipeline}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marcus Wagner \email{marcus.wagner@tum.de} (\href{https://orcid.org/0009-0000-9421-6766}{ORCID})

Authors:
\itemize{
  \item Marcus Wagner \email{marcus.wagner@tum.de} (\href{https://orcid.org/0009-0000-9421-6766}{ORCID})
}

}
\keyword{internal}
